touch up mq
