+++ swagger_server/controllers/topology_controller.py
@@ -2,18 +2,20 @@ import connexion
 import six
 import os
 import json
+import logging
 
 from swagger_server.models.api_response import ApiResponse  # noqa: E501
 from swagger_server.models.topology import Topology  # noqa: E501
 from swagger_server import util
 from swagger_server.utils.db_utils import *
-from swagger_server.messaging.message_queue_consumer import *
-from swagger_server.messaging.rpc_queue_consumer import *
-from swagger_server.messaging.async_publisher import *
+from swagger_server.messaging.rpc_queue_producer import *
+# from swagger_server.messaging.async_publisher import *
+# from swagger_server.messaging.message_queue_consumer import *
 
 LOG_FORMAT = ('%(levelname) -10s %(asctime)s %(name) -30s %(funcName) '
               '-35s %(lineno) -5d: %(message)s')
 LOGGER = logging.getLogger(__name__)
+logging.getLogger("pika").setLevel(logging.WARNING)
 
 class Payload(object):
     def __init__(self, j):
@@ -28,12 +30,7 @@ db_tuples = [('config_table', "test-config")]
 db_instance = DbUtils()
 db_instance._initialize_db(DB_NAME, db_tuples)
 
-# rpc = RpcClient()
-
-publisher = AsyncPublisher(
-        'amqp://guest:guest@aw-sdx-monitor.renci.org:5672/%2F?connection_attempts=3&heartbeat=3600'
-    )
-# publisher.publish_message('header', 'message')
+rpc = RpcClient()
 
 def add_topology(body):  # noqa: E501
     """Send a new topology to SDX-LC
@@ -55,15 +52,14 @@ def add_topology(body):  # noqa: E501
     db_instance.add_key_value_pair_to_db('test', json_body)
     print('Saving to database complete.')
 
+
     print("Published Message: {}".format(body))
     # response = publisher.call(json_body)
-    publisher.run()
-    print(" [.] Got response: " + str(response))
+    # message_number = publisher.publish_message()
+    response = rpc.call(json_body)
+    print('Published message.')
 
-
-    return 'do some magic!'
-    # return str(response)
-
+    return str(response)
 
 def delete_topology(topology_id, api_key=None):  # noqa: E501
     """Deletes a topology
