+++ swagger_server/messaging/message_queue_producer.py
@@ -8,7 +8,7 @@ class MetaClass(type):
             return cls._instance[cls]
 
 class RabbitmqConfigure(metaclass=MetaClass):
-    def __init__(self, queue='hello', host='localhost', routingKey='hello', exchange=''):
+    def __init__(self, queue='hello', host='aw-sdx-monitor.renci.org', routingKey='hello', exchange=''):
         #Configure Rabbit Mq Server
         self.queue = queue
         self.host = host
@@ -33,3 +33,16 @@ class RabbitMq():
         print("Published Message: {}".format(body))
         
         self._connection.close()
+
+if __name__ == "__main__":
+
+    server = RabbitmqConfigure(queue='hello',
+                               host='localhost',
+                               routingKey='hello',
+                               exchange='')
+
+    image = Image(filename="local-ctlr-1.manifest")
+    data = image.get
+
+    with RabbitMq(server) as rabbitmq:
+        rabbitmq.publish(payload=data)
