+++ swagger_server/messaging/rpc_queue_consumer.py
@@ -1,38 +1,25 @@
 #!/usr/bin/env python
 import pika
-import uuid
-import os
 
-MQ_HOST = os.environ.get('MQ_HOST')
+def on_request(ch, method, props, body):
+    response = body
 
-class RpcClient(object):
-    def __init__(self):
-        self.connection = pika.BlockingConnection(
-            pika.ConnectionParameters(host=MQ_HOST))
+    ch.basic_publish(exchange='',
+                     routing_key=props.reply_to,
+                     properties=pika.BasicProperties(correlation_id = \
+                                                         props.correlation_id),
+                     body=str(response))
+    ch.basic_ack(delivery_tag=method.delivery_tag)
 
-        self.channel = self.connection.channel()
+if __name__ == "__main__":
+    connection = pika.BlockingConnection(pika.ConnectionParameters(host='aw-sdx-monitor.renci.org'))
 
-        result = self.channel.queue_declare(queue='', exclusive=True)
-        self.callback_queue = result.method.queue
+    channel = connection.channel()
 
-        self.channel.basic_consume(queue=self.callback_queue,
-                            on_message_callback=self.on_response,
-                            auto_ack=True)
+    channel.queue_declare(queue='rpc_queue')
 
-    def on_response(self, ch, method, props, body):
-        if self.corr_id == props.correlation_id:
-            self.response = body
+    channel.basic_qos(prefetch_count=1)
+    channel.basic_consume(queue='rpc_queue', on_message_callback=on_request)
 
-    def call(self, body):
-        self.response = None
-        self.corr_id = str(uuid.uuid4())
-        self.channel.basic_publish(exchange='',
-                                routing_key='rpc_queue',
-                                properties=pika.BasicProperties(
-                                    reply_to=self.callback_queue,
-                                    correlation_id=self.corr_id,
-                                ),
-                                body=str(body))
-        while self.response is None:
-            self.connection.process_data_events()
-        return self.response
+    print("Awaiting RPC requests")
+    channel.start_consuming()
